requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
httpx>=0.27.0
